# Optimize the logic to compare the routing entries between ASIC_DB and the prefixes pushed to DUT