@stephenxs 2022-06-07
test_bgp_update_timer failed
```
2022-06-07T01:54:37.4959405Z =========================== short test summary info ============================
2022-06-07T01:54:37.4960197Z FAILED bgp/test_bgp_update_timer.py::test_bgp_update_timer[vlab-01] - RunAnsi...
2022-06-07T01:54:37.4960782Z ========================== 1 failed in 106.10 seconds ==========================
2022-06-07T01:54:37.4961303Z INFO:root:Can not get Allure report URL. Please check logs
```