@ZhaohuiS 2022-06-18
@stephenxs How do we see the following improvement from your log?
IPv4: 20 minutes => 20 seconds
IPv6: 22 minutes => 2 minutes