@ZhaohuiS 2022-06-20
> > 
> 
> Hi @ZhaohuiS For IPv4, the optimized time should be the difference between time stamps of the following message, ~15s
> 
> ```
> 2022-06-06 07:26:50 test_route_perf.exec_routes              L0203 INFO   | All route entries have been installed in ASIC_DB in 66.122285 seconds
> 2022-06-06 07:27:04 test_route_perf.test_perf_add_remove_rou L0253 INFO   | Time to set 100000 ipv4 routes is 66.12 seconds.
> ```
> 
> There is no too much log message in the original script but the time before optimizing can be estimated as `end-2-end-time` - `time used to push to ASIC_DB`, which is `21 minutes 36 seconds` - `~70 seconds` => `20 minutes 26 seconds`
> 
> ```
> 2022-06-06 03:50:10 test_route_perf.test_perf_add_remove_rou L0229 INFO   | IP route utilization before test start: Used: 6414, Available: 186352, Test count: 100000
> 2022-06-06 04:08:05 test_route_perf.test_perf_add_remove_rou L0244 INFO   | Time to set 100000 ipv4 routes is 69.45 seconds.
> 2022-06-06 04:12:46 test_route_perf.test_perf_add_remove_rou L0248 INFO   | Time to del 100000 ipv4 routes is 50.76 seconds.
> ```
> 
> IPv6 is the same.

@stephenxs Thank you for your clarification. I just saw "Time to set 100000 ipv4 routes is 66.12 seconds." This number is similar. But yes, as you said, the log timestamp is huge different.
Appreciate your effort on improving this, it saves a lot of running time.